# ffs-api/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API for user authentication, creation, and management of schools, districts, bids, and tasks"
  },
  "paths": {
...
    "/api/schools/{districtId}/{id}": {
      "get": {
        "summary": "Get school details by ID",
        "description": "Retrieve detailed information about a specific school within a district by its ID.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "The ID of the district containing the school",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "schoolId",
            "in": "path",
            "required": true,
            "description": "The ID of the school to retrieve",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchoolByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid district ID or school ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found - School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a school",
        "description": "Update an existing school's details within a district.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district the school belongs to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the school to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "School updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSchoolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft delete a school",
        "description": "Soft delete a school with the specified ID within a district.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district the school belongs to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the school to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "summary": "Get tasks assigned to authenticated user",
        "description": "Retrieve a list of tasks assigned to the authenticated user.",
        "tags": ["Tasks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks assigned to the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Create a new task with title, due date, assigned user, and completion status.",
        "tags": ["Tasks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "patch": {
        "summary": "Update a task",
        "description": "Update selected fields of an existing task, supporting partial updates for title, due date, assigned user, and completion status.",
        "tags": ["Tasks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the task to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid task ID or request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Delete a task with the specified ID. User can only delete tasks they own.",
        "tags": ["Tasks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the task to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully (No Content)"
          },
          "400": {
            "description": "Invalid task ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not own this task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateSchoolRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the school",
            "minLength": 1
          },
          "enrollment": {
            "type": "integer",
            "description": "Number of students enrolled"
          },
          "school_type": {
            "type": "string",
            "enum": ["High School", "Middle School", "Elementary School", "Childcare"],
            "description": "Type of school"
          },
          "address_line_1": {
            "type": "string",
            "description": "First line of school address"
          },
          "address_line_2": {
            "type": "string",
            "description": "Second line of school address"
          },
          "city": {
            "type": "string",
            "description": "City where the school is located"
          },
          "state": {
            "type": "string",
            "description": "State where the school is located"
          },
          "zip_code": {
            "type": "string",
            "description": "ZIP code of the school"
          },
          "shipping_address_line_1": {
            "type": "string",
            "description": "First line of shipping address"
          },
          "shipping_address_line_2": {
            "type": "string",
            "description": "Second line of shipping address"
          },
          "shipping_address_city": {
            "type": "string",
            "description": "City of shipping address"
          },
          "shipping_address_state": {
            "type": "string",
            "description": "State of shipping address"
          },
          "shipping_address_zip_code": {
            "type": "string",
            "description": "ZIP code of shipping address"
          },
          "school_contact_name": {
            "type": "string",
            "description": "Name of school contact"
          },
          "school_contact_phone": {
            "type": "string",
            "description": "Phone number of school contact"
          },
          "school_contact_email": {
            "type": "string",
            "format": "email",
            "description": "Email address of school contact"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about the school"
          },
          "override_district_billing": {
            "type": "boolean",
            "description": "Whether to override district billing settings"
          },
          "contact_first_name": {
            "type": "string",
            "description": "First name of primary contact"
          },
          "contact_last_name": {
            "type": "string",
            "description": "Last name of primary contact"
          },
          "contact_title": {
            "type": "string",
            "description": "Title of primary contact"
          },
          "contact_phone": {
            "type": "string",
            "description": "Phone number of primary contact"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "description": "Email address of primary contact"
          },
          "billing_contact": {
            "type": "string",
            "description": "Full name of billing contact"
          },
          "billing_phone": {
            "type": "string",
            "description": "Phone number of billing contact"
          },
          "billing_email": {
            "type": "string",
            "format": "email",
            "description": "Email address of billing contact"
          },
          "billing_address_line_1": {
            "type": "string",
            "description": "First line of billing address"
          },
          "billing_address_line_2": {
            "type": "string",
            "description": "Second line of billing address"
          },
          "billing_city": {
            "type": "string",
            "description": "City of billing address"
          },
          "billing_state": {
            "type": "string",
            "description": "State of billing address"
          },
          "billing_zip_code": {
            "type": "string",
            "description": "ZIP code of billing address"
          },
          "fax": {
            "type": "string",
            "description": "School fax number"
          },
          "shippingInstructions": {
            "type": "string",
            "description": "Special shipping instructions for the school"
          },
          "shippingDeliveryHours": {
            "type": "string",
            "description": "Preferred delivery hours for shipments to the school"
          }
        },
        "description": "Request body for updating a school. All fields are optional to support partial updates. At least one field must be provided."
      },
      "UpdateSchoolResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the updated school"
          },
          "name": {
            "type": "string",
            "description": "Name of the school"
          },
          "district_id": {
            "type": "integer",
            "description": "ID of the district the school belongs to"
          }
        },
        "required": ["id", "name", "district_id"]
      },
...
      "CreateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the task"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "description": "Due date for the task"
          },
          "assigned_to": {
            "type": "integer",
            "description": "ID of the user assigned to the task"
          },
          "completed": {
            "type": "boolean",
            "description": "Completion status of the task"
          }
        },
        "required": ["title", "assigned_to"]
      },
      "CreateTaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the created task"
          },
          "title": {
            "type": "string",
            "description": "Title of the task"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "description": "Due date for the task"
          },
          "assigned_to": {
            "type": "integer",
            "description": "ID of the user assigned to the task"
          },
          "completed": {
            "type": "boolean",
            "description": "Completion status of the task"
          }
        },
        "required": ["id", "title", "assigned_to"]
      },
...
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the task"
          },
          "title": {
            "type": "string",
            "description": "Title of the task"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "description": "Due date for the task"
          },
          "assigned_to": {
            "type": "integer",
            "description": "ID of the user assigned to the task"
          },
          "completed": {
            "type": "boolean",
            "description": "Completion status of the task"
          }
        },
        "required": ["id", "title", "assigned_to"]
      },
...
      "GetSchoolByIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the school"
          },
          "districtId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the district this school belongs to"
          },
          "name": {
            "type": "string",
            "description": "Name of the school"
          },
          "enrollment": {
            "type": "integer",
            "nullable": true,
            "description": "Number of students enrolled"
          },
          "schoolType": {
            "type": "string",
            "description": "Type of school (e.g., Elementary, Middle, High)"
          },
          "addressLine1": {
            "type": "string",
            "nullable": true,
            "description": "First line of the school address"
          },
          "addressLine2": {
            "type": "string",
            "nullable": true,
            "description": "Second line of the school address"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "description": "City where the school is located"
          },
          "state": {
            "type": "string",
            "nullable": true,
            "description": "State where the school is located"
          },
          "zipCode": {
            "type": "string",
            "nullable": true,
            "description": "ZIP code of the school"
          },
          "shippingAddressLine1": {
            "type": "string",
            "nullable": true,
            "description": "First line of the shipping address"
          },
          "shippingAddressLine2": {
            "type": "string",
            "nullable": true,
            "description": "Second line of the shipping address"
          },
          "shippingAddressCity": {
            "type": "string",
            "nullable": true,
            "description": "City for shipping address"
          },
          "shippingAddressState": {
            "type": "string",
            "nullable": true,
            "description": "State for shipping address"
          },
          "shippingAddressZipCode": {
            "type": "string",
            "nullable": true,
            "description": "ZIP code for shipping address"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes about the school"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "School phone number"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "School email address"
          },
          "overrideDistrictBilling": {
            "type": "boolean",
            "description": "Whether this school overrides district billing settings"
          },
          "status": {
            "type": "string",
            "description": "Current status of the school"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the school record was created"
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "School code identifier"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Location description"
          },
          "directorName": {
            "type": "string",
            "nullable": true,
            "description": "Name of the school director"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "description": "School website URL"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Description of the school"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "URL to school logo"
          },
          "fullAddress": {
            "type": "string",
            "description": "Complete formatted address"
          },
          "shippingFullAddress": {
            "type": "string",
            "description": "Complete formatted shipping address"
          }
        },
        "required": ["id", "districtId", "name", "schoolType", "overrideDistrictBilling", "status", "createdAt", "fullAddress", "shippingFullAddress"]
      },
...
    }
  }
}